restructure py files for easier coding and reviewing
